+++ src/Layers/xrRenderDX10/dx10ConstantBuffer_impl.h
@@ -172,14 +172,14 @@ IC void dx10ConstantBuffer::seta(R_constant* C, R_constant_load& L, u32 e, const
     // c_f.dirty	(base,base+1);
 }
 
-IC void* dx10ConstantBuffer::AccessDirect(R_constant_load& L, u32 DataSize)
+IC void* dx10ConstantBuffer::AccessDirect(R_constant_load& L, size_t DataSize)
 {
     //	Check buffer size in client code: don't know if actual data will cross
     //	buffer boundaries.
     VERIFY(L.index < (int)m_uiBufferSize);
     BYTE* res = ((BYTE*)m_pBufferData) + L.index;
 
-    if ((u32)L.index + DataSize <= m_uiBufferSize)
+    if ((size_t)L.index + DataSize <= m_uiBufferSize)
     {
         m_bChanged = true;
         return res;
